Update Ch02.md
Line 16: I think you meant "in-product" and was confident enough to make the change, but I thought I'd flag it in case it was a specific term you use for RavenDB. If it is, let me know so I don't change it in future instances.

Line 33: I'm pretty sure that you mean the word "listing" to refer to your in-book code snippets--like an alternative to figure names--but I had to do some research to figure that out. More importantly, Erik wasn't familiar with this naming convention, so it may confuse even techies who deal with code snippets a lot in their literary travels. It could theoretically be read it as "The output of this command should look something like, 'Listing 2.1,'" as if "Listing 2.1" was the verbatim command output. You might want to caption your snippets to avoid confusion. You could also say "Listing 2.1 below" in the sentence. It could still be misread, but it's less likely.

Lines 74 and 94: Erik had the following to say: "I would suggest adding a sentence clarifying that you should repeat until you find an open port or something.  (i.e. it's possible that something is also running on 8081)"

Line 88: Just making sure you don't forget your to-do list!

Line 116: I'm not sure what the caret means in html, and Google and Erik aren't any help here. In the GitHub preview, it doesn't render as a footnote, but I suspect that's what you're doing. If that's not the case and you mean to include the text in brackets as an in-text parenthetical, I suggest replacing your current text with the following: 

What we have right now is a RavenDB node that is a self-contained cluster. (Actually, that's not exactly the case, but the details on the state of a newly minted node are a bit complex and covered in more detailed in the "Your First RavenDB Cluster" chapter.) 

Line 144: Erik and I both think you need to work on this sentence:

Every document belongs to exactly one collection, usually based on the entity type of the document on the client

I'm having trouble specifically understanding your prepositions in the last phrases ("based on the entity type *of* the document *on* the client") and I'm not in a good position to suggest a change. Erik wants you to clarify who the client is, since this is the first time one is mentioned.

Line 149: You've got "storage" misspelled, but I'm not sure if it's integral to where you're linking

Line 172: Alternate correction option, if you prefer it: "Though it's pretty, I'll be the first to admit that looking at a syntax highlighted text editor isn't really that impressive." 

Line 179: Double-checking that the em dash (long line separating "next section" and "no worries) will render in whatever your final output is. It works in the preview, but I'm not sure what you'll use in the end to publish.