Update Ch02.md
Line 472-73: I noticed a break in the preview mode that looks awkward, and I'm not sure what's causing it or if it even matters in the final output.

Line 810: I added a phrase, since the transition to the next sentence made it sound like you were excited about the extra trip to the database. Reword as you see fit, but I think it's worth introducing the feature by name here.  

Line 885: Your use of "we" can be confusing, since sometimes you use it to talk as if you and a reader are on a journey learning together ("We'll type the following code"). But you also use "we" to refer to your team that created the database. This isn't a problem when it's clear which mode you're in--I rather like it--but in situations like in line 885, I'm not sure if you mean "the choices we'll make as you and I are learning together at later point in the book" or "the choices we made as developers while writing this program." That the difference between "we'll make" and "we made," and you can choose which you meant based on this feedback. 

--Quick note--I forgot to say this earlier, but I decapitalized all instances of "just work" and italicized them instead, and I wanted to let you know why. People sometimes do capitalize Big Ideas, but it's often confusing to a reader--it looks like a mistake, or it makes people wonder if you're talking about a company or something else that calls for proper noun treatment. I don't want anyone being distracted from your text, so I figure italicizing will do what you want without distracting.