Update Ch02.md
Line 1129: A note on this change--I think your footnote here (and the story about emojis breaking the build) is hilarious. But I'd be a bad editor if I didn't draw your attention to things that have the possibility of some culturally/socially-charged blowback. If you specify that they're teenage *girls*,  people who feel strongly about the intersection of women and tech--a passionate and vocal crew--might make this a PR nightmare. I changed it to "teenage developers" because it eliminates that social risk, and you get to keep your amusing side note without losing much meaning.